Split src/ Types.ts into d.ts and Constants
