+++ src/renderer/Types.d.ts
@@ -8,19 +8,6 @@ import { IDisposable } from 'xterm';
 import { IColorSet } from 'browser/Types';
 import { IRenderDimensions, CharacterJoinerHandler } from 'browser/renderer/Types';
 
-/**
- * Flags used to render terminal text properly.
- */
-export const enum FLAGS {
-  BOLD = 1,
-  UNDERLINE = 2,
-  BLINK = 4,
-  INVERSE = 8,
-  INVISIBLE = 16,
-  DIM = 32,
-  ITALIC = 64
-}
-
 export interface IRenderLayer extends IDisposable {
   /**
    * Called when the terminal loses focus.
